tododo-api/pyproject.toml
[tool.poetry]
name = "tododo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.2"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
gunicorn = "^20.1.0"
fastapi-csrf-protect = "^0.3.0"
pyjwt = "^2.7.0"
passlib = "^1.7.4"
dnspython = "^2.3.0"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
